Refactored code to remove duplications. Moved DOMParser configuration.
